+++ docs/docs.go
@@ -6006,6 +6006,9 @@ var doc = `{
                 "date": {
                     "type": "string"
                 },
+                "date_created": {
+                    "type": "string"
+                },
                 "encrypted_blob": {
                     "type": "string"
                 },
@@ -7504,6 +7507,9 @@ var doc = `{
                 "status"
             ],
             "properties": {
+                "date": {
+                    "type": "string"
+                },
                 "encrypted_blob": {
                     "type": "string"
                 },
@@ -8136,7 +8142,7 @@ type swaggerInfo struct {
 
 // SwaggerInfo holds exported Swagger Info so clients can modify it
 var SwaggerInfo = swaggerInfo{
-	Version:     "2.4.0",
+	Version:     "2.5.0",
 	Host:        "localhost",
 	BasePath:    "/health",
 	Schemes:     []string{"https"},