+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.5.0] - 2021-01-20
+### Changed
+- Update the process manual test admin API. [#114](https://github.com/rokwire/health-building-block/issues/114)
+
 ## [2.4.0] - 2021-01-11
 ### Security
 - Prepare the APIs to be consumed by web app. [#111](https://github.com/rokwire/health-building-block/issues/111)
@@ -209,7 +213,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update access rule admin api.
 - Delete access rule admin api.
 
-[Unreleased]: https://github.com/rokwire/health-building-block/compare/v2.4.0...HEAD
+[Unreleased]: https://github.com/rokwire/health-building-block/compare/v2.5.0...HEAD
+[2.5.0]: https://github.com/rokwire/health-building-block/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/rokwire/health-building-block/compare/v2.3.0...v2.4.0
 [2.3.0]: https://github.com/rokwire/health-building-block/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/rokwire/health-building-block/compare/v2.1.0...v2.2.0